MISC: mkassets.sh: always use /tmp/anklang-mkassets/ for release builds

+++ misc/mkassets.sh
@@ -1,68 +1,67 @@
 #!/bin/bash
 # This Source Code Form is licensed MPL-2.0: http://mozilla.org/MPL/2.0
-set -Eeuo pipefail #-x
+set -Eeuo pipefail && SCRIPTNAME=${0##*/} && die() { [ -z "$*" ] || echo "$SCRIPTNAME: $*" >&2; exit 127 ; }
 
-SCRIPTNAME=`basename $0`
-function die  { [ -n "$*" ] && echo "$SCRIPTNAME: $*" >&2; exit 127 ; }
+# Usage: mkassets.sh <TARDIR>/<TARBALL>
+# Build assets from TARBALL in a separate directory and copy those assets
+# together with `build-assets` into TARDIR.
+
 umask 022	# create 0755 dirs by default
-ORIG_PWD=$PWD
 
 # paths & commands
 TARBALL=`readlink -f $1`
 TARDIR=`dirname $TARBALL`
 ASSETLIST=$TARDIR/build-assets
-UID_=`id -u`
-OOTBUILD="${OOTBUILD:-/tmp/anklang-ootbuild$UID_}"
+SCRATCHDIR="/tmp/anklang-mkassets/"
 MAKE="make -w V=${V:-}"
 
-# prepare build dir
-mkdir -p $OOTBUILD/
-( cd $OOTBUILD
+# Clear build dir, but keep it in place in case it is a mount point
+mkdir -p $SCRATCHDIR/
+test -O $SCRATCHDIR -a -w $SCRATCHDIR || die "SCRATCHDIR: build directory not writable"
+( cd $SCRATCHDIR
   rm -rf * .[^.]* ..?*
 )
 
-# link .dlcache
+# Link .dlcache to reduce network IO
 test -d .dlcache/ && {
-  rm -rf $OOTBUILD/.dlcache
-  ln -s $PWD/.dlcache $OOTBUILD/.dlcache
+  rm -rf $SCRATCHDIR/.dlcache
+  ln -s $PWD/.dlcache $SCRATCHDIR/.dlcache
 }
 
 # extract tarball, extract release infos
-tar xf $TARBALL -C $OOTBUILD --strip-components=1
-( cd $OOTBUILD
+tar xf $TARBALL -C $SCRATCHDIR --strip-components=1
+( cd $SCRATCHDIR
   misc/version.sh
 )		> $TARDIR/build-version
-cp $OOTBUILD/NEWS.md $TARDIR/build-news
-OOTBUILD_VERSION=`cut '-d ' -f1 $TARDIR/build-version`
+cp $SCRATCHDIR/NEWS.md $TARDIR/build-news
+TAR_VERSION=`cut '-d ' -f1 $TARDIR/build-version`
 
-# build and check
-( cd $OOTBUILD
+# Change to $SCRATCHDIR to build various assets
+( cd $SCRATCHDIR
+
+  # Build and check, a `production+sse` build also triggers FMA builds in ui/Makefile.mk
   $MAKE -j`nproc` \
 	MODE=production \
 	INSN=sse \
 	all pdf
   $MAKE check
-)
-cp $OOTBUILD/out/doc/anklang-manual.pdf $OOTBUILD/out/anklang-manual-$OOTBUILD_VERSION.pdf
-cp $OOTBUILD/out/doc/anklang-internals.pdf $OOTBUILD/out/anklang-internals-$OOTBUILD_VERSION.pdf
+  cp out/doc/anklang-manual.pdf out/anklang-manual-$TAR_VERSION.pdf
+  cp out/doc/anklang-internals.pdf out/anklang-internals-$TAR_VERSION.pdf
 
-# make Deb
-( cd $OOTBUILD
+  # Make Deb
   BUILDDIR=out V=$V misc/mkdeb.sh
-)
 
-# make AppImage
-( cd $OOTBUILD
+  # Make AppImage
   BUILDDIR=out V=$V misc/mkAppImage.sh
-  test ! -r /dev/fuse || time out/anklang-$OOTBUILD_VERSION-x64.AppImage --quitstartup
+  test ! -r /dev/fuse || time out/anklang-$TAR_VERSION-x64.AppImage --quitstartup
 )
 
 # Gather assets
 echo ${TARDIR##*/}/${TARBALL##*/}	>  $ASSETLIST.tmp
-for f in $OOTBUILD/out/anklang_${OOTBUILD_VERSION}_amd64.deb \
-	 $OOTBUILD/out/anklang-$OOTBUILD_VERSION-x64.AppImage \
-	 $OOTBUILD/out/anklang-manual-$OOTBUILD_VERSION.pdf \
-	 $OOTBUILD/out/anklang-internals-$OOTBUILD_VERSION.pdf \
+for f in $SCRATCHDIR/out/anklang_${TAR_VERSION}_amd64.deb \
+	 $SCRATCHDIR/out/anklang-$TAR_VERSION-x64.AppImage \
+	 $SCRATCHDIR/out/anklang-manual-$TAR_VERSION.pdf \
+	 $SCRATCHDIR/out/anklang-internals-$TAR_VERSION.pdf \
 	 ; do
   cp $f $TARDIR
   echo ${TARDIR##*/}/${f##*/}		>> $ASSETLIST.tmp
